backend/requirements.txt
flask
flask-cors
tensorflow
numpy
pandas
opencv-python
pillow
scikit-learn